+++ api/app/detector/vulnerability_detector.py
@@ -1,8 +1,9 @@
 from sqlalchemy.orm import Session
 
 from app import models, persistence
-from app.detector import version
-from app.detector.version import PackageFamily, VulnerableRange
+from app.detector.package_family import PackageFamily
+from app.detector.version import version_factory
+from app.detector.vulnerable_range import VulnerableRange
 
 
 def detect_vulnerability_by_dependency(db: Session, dependency: models.Dependency) -> list[str]:
@@ -41,7 +42,9 @@ def detect_vulnerability_by_affect(affect: models.Affect) -> set[str]:
 
 def _detect_matched(package_version: models.PackageVersion, affect: models.Affect) -> bool:
     package_family = PackageFamily.from_registry(package_version.package.ecosystem)
-    dependency_version = version.gen_version_instance(package_family, package_version.version)
+    dependency_version = version_factory.gen_version_instance(
+        package_family, package_version.version
+    )
 
     for affected_version in _get_affected_versions(affect):
         try:
